refactor: Centralize connected realm scheduling
Moves the `@Scheduled` annotation and concurrent execution guard from `ConnectedRealmService` to a new `ConnectedRealmSchedule` component. This separates scheduling concerns from the core business logic of updating realms, improving modularity and clarity of responsibility. The `ConnectedRealmSchedule` now manages both startup synchronization and recurring updates.
