+++ backend/src/test/kotlin/net/jonasmf/auctionengine/schedules/ConnectedRealmScheduleTest.kt
@@ -1,4 +1,4 @@
-package net.jonasmf.auctionengine.service
+package net.jonasmf.auctionengine.schedules
 
 import ch.qos.logback.classic.Logger
 import ch.qos.logback.classic.spi.ILoggingEvent
@@ -6,13 +6,15 @@ import ch.qos.logback.core.read.ListAppender
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import org.junit.jupiter.api.Assertions.assertTrue
+import net.jonasmf.auctionengine.service.AuthService
+import net.jonasmf.auctionengine.service.ConnectedRealmService
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.slf4j.LoggerFactory
 import org.springframework.boot.DefaultApplicationArguments
 import reactor.core.publisher.Mono
 
-class ConnectedRealmStartupSyncTest {
+class ConnectedRealmScheduleTest {
     @Test
     fun `run triggers connected realm update on startup`() {
         val authService = mockk<AuthService>()
@@ -20,7 +22,7 @@ class ConnectedRealmStartupSyncTest {
         every { authService.ensureToken() } returns Mono.just("token")
         every { connectedRealmService.updateRealms() } returns Unit
 
-        ConnectedRealmStartupSync(authService, connectedRealmService).run(DefaultApplicationArguments())
+        ConnectedRealmSchedule(authService, connectedRealmService).run(DefaultApplicationArguments())
 
         verify(exactly = 1) { authService.ensureToken() }
         verify(exactly = 1) { connectedRealmService.updateRealms() }
@@ -31,15 +33,15 @@ class ConnectedRealmStartupSyncTest {
         val authService = mockk<AuthService>()
         val connectedRealmService = mockk<ConnectedRealmService>()
         val listAppender = attachAppender()
-        every { authService.ensureToken() } throws RuntimeException("boom")
+        every { authService.ensureToken() } throws RuntimeException("failure")
 
         try {
-            ConnectedRealmStartupSync(authService, connectedRealmService).run(DefaultApplicationArguments())
+            ConnectedRealmSchedule(authService, connectedRealmService).run(DefaultApplicationArguments())
 
             val messages = listAppender.list.map(ILoggingEvent::getFormattedMessage)
-            assertTrue(
+            Assertions.assertTrue(
                 messages.any {
-                    it.contains("Connected realm startup synchronization failed. Continuing application startup.")
+                    it.contains("Connected realm startup synchronization failed.")
                 },
             )
             verify(exactly = 1) { authService.ensureToken() }
@@ -50,7 +52,7 @@ class ConnectedRealmStartupSyncTest {
     }
 
     private fun attachAppender(): ListAppender<ILoggingEvent> {
-        val logger = LoggerFactory.getLogger(ConnectedRealmStartupSync::class.java) as Logger
+        val logger = LoggerFactory.getLogger(ConnectedRealmSchedule::class.java) as Logger
         return ListAppender<ILoggingEvent>().also {
             it.start()
             logger.addAppender(it)
@@ -58,7 +60,7 @@ class ConnectedRealmStartupSyncTest {
     }
 
     private fun detachAppender(appender: ListAppender<ILoggingEvent>) {
-        val logger = LoggerFactory.getLogger(ConnectedRealmStartupSync::class.java) as Logger
+        val logger = LoggerFactory.getLogger(ConnectedRealmSchedule::class.java) as Logger
         logger.detachAppender(appender)
         appender.stop()
     }
